% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{betaStr}
\alias{betaStr}
\title{Constructs an APA formatted string for an estimated regression coefficient.}
\usage{
betaStr(param, modObj, ...)
}
\arguments{
\item{param}{The parameter of interest. Can be "int" as a shortcut 
for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for an estimated regression coefficient.
}
