% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{confIntStr}
\alias{confIntStr}
\title{Constructs an APA formatted string for a confidence interval around
a parameter estimate from a model.}
\usage{
confIntStr(param, modObj, digits = NULL, confLvl = NULL, asEqn = NULL)
}
\arguments{
\item{param}{The parameter of interest. Can be "int" as a shortcut 
for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}

\item{digits}{Number of digits to round to.}

\item{confLvl}{The alpha for confidence interval estimation.}

\item{asEqn}{Should the string be an equation?}
}
\description{
Constructs an APA formatted string for a confidence interval around
a parameter estimate from a model.
}
