% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{confIntSum}
\alias{confIntSum}
\title{Confidence interval for summary objects. Derived directly from 
\code{confint.lm()}.}
\usage{
confIntSum(object, parm, confLvl = NULL)
}
\arguments{
\item{object}{Summary object of lm model.}

\item{parm}{a specification of which parameters are to be given confidence 
intervals, either a vector of numbers or a vector of names. If missing, 
all parameters are considered.}

\item{confLvl}{Alpha value.}
}
\description{
Confidence interval for summary objects. Derived directly from 
\code{confint.lm()}.
}
