% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{eta2StrFromLm}
\alias{eta2StrFromLm}
\title{Constructs an APA formatted string for an \eqn{\eta^2_p} value. Note that
this needs to be an equation for the character to display correctly in 
Markdown (\eqn{LaTeX}).}
\usage{
eta2StrFromLm(param, modObj, ...)
}
\arguments{
\item{param}{The parameter of interest. Can be "total" for the omnibus
f-value, or "int" as a shortcut for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for an \eqn{\eta^2_p} value. Note that
this needs to be an equation for the character to display correctly in 
Markdown (\eqn{LaTeX}).
}
