% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{eta2StrFromModelCompare}
\alias{eta2StrFromModelCompare}
\title{Constructs an APA formatted string for an \eqn{\eta^2_p} value. Note that
this needs to be an equation for the character to display correctly in 
Markdown (\eqn{LaTeX}).}
\usage{
eta2StrFromModelCompare(mcObj, ...)
}
\arguments{
\item{mcObj}{Output from \code{modelCompare}}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for an \eqn{\eta^2_p} value. Note that
this needs to be an equation for the character to display correctly in 
Markdown (\eqn{LaTeX}).
}
