% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{fStr}
\alias{fStr}
\title{Constructs an APA formatted string for an F value.}
\usage{
fStr(param, modObj, digits = NULL, asEqn = NULL, dfDigits = NULL)
}
\arguments{
\item{param}{The parameter of interest. Can be "total" for the omnibus
f-value, or "int" as a shortcut for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for an F value.
}
