% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{fStrFromModelCompare}
\alias{fStrFromModelCompare}
\title{Constructs an APA formatted string for an F value from a model comparison.}
\usage{
fStrFromModelCompare(mcObj, ...)
}
\arguments{
\item{mcObj}{The parameter of interest. Can be "total" for the omnibus
f-value, or "int" as a shortcut for "(Intercept)".}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for an F value from a model comparison.
}
