% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{getOpts}
\alias{getOpts}
\title{A way to quickly get options for functions.}
\usage{
getOpts(arguments = NULL, ...)
}
\arguments{
\item{arguments}{A character vector of arguments to access. If NULL,
the function processes all available arguments.}

\item{...}{Arguments that override the defaults. If an argument is provided
via ellipses from a caller function, these are assigned the provided value 
instead of the default value.}
}
\description{
A way to quickly get options for functions.
}
\details{
If the
}
