% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{mStr}
\alias{mStr}
\title{Prints an APA formatted mean.}
\usage{
mStr(v, na.rm = F, digits = NULL, asEqn = NULL)
}
\arguments{
\item{v}{Either a mean value or a vector. If a vector, then \code{mean()} 
is called on it.}

\item{na.rm}{If v is a vector, this is passed to \code{mean()}. 
Default: FALSE.}
}
\description{
Prints an APA formatted mean.
}
