% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{printStats}
\alias{printStats}
\title{Print a Formatted Stats String}
\usage{
printStats(param, modObj, ...)
}
\arguments{
\item{param}{The parameter of interest. The shortcut "int" is recognized.}

\item{modObj}{The \code{lm} or \code{summary.lm} object.}

\item{...}{Other options.}

\item{template}{A string specifying which stats to include in this string. 
See details for how to construct a string. If left as \code{NULL}, the 
default template is used.}
}
