% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{printStatsFromModelCompare}
\alias{printStatsFromModelCompare}
\title{Print a Formatted Stats String from a \code{modelCompare} Object}
\usage{
printStatsFromModelCompare(mcObj, asCor = F, ...)
}
\arguments{
\item{mcObj}{The object resulting from a \code{modelCompare} call.}

\item{...}{Other options.}

\item{template}{A string specifying which stats to include in this string. 
See details for how to construct a string. If left as \code{NULL}, the 
default template is used.}
}
\description{
Print a Formatted Stats String from a \code{modelCompare} Object
}
\details{
The \code{template} argument is a string containing a sequence of the
following characters, each separated by an underscore:
\itemize{
  \item f: An 1-degree-of-freedom F-statistic, e.g., \eqn{F(1,36)=4.2}
  \item t: A t-statistic, e.g., \eqn{t(38)=2.1}
  \item p: A p-value, e.g., \eqn{p=.03} or \eqn{p<.01} depending on the 
  value
  \item eta2: An \eqn{\eta^2_p} value, e.g., \eqn{\eta^2_p=.056}
  \item deltaR2: not currently implemented
  \item ci: A confidence interval, e.g., \eqn{95\%CI[.01, 2.34]}. Not 
  currently implemented.
}
Note that the same set of \code{...} arguments is passed to every function 
call. Otherwise, the strings are built with default argument values as
normal.
}
