% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{pStrFromModelCompare}
\alias{pStrFromModelCompare}
\title{Constructs a string for a p-value from \code{modelCompare} output.}
\usage{
pStrFromModelCompare(mcObj, ...)
}
\arguments{
\item{mcObj}{An object from \code{modelCompare}.}

\item{...}{Options.}
}
\description{
Constructs a string for a p-value from \code{modelCompare} output.
}
