% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{resHelpPrep}
\alias{resHelpPrep}
\title{Converts a model object to a summary object if necessary, validates
the parameter information, and grabs options if necessary.}
\usage{
resHelpPrep(param, modObj)
}
\arguments{
\item{param}{A string containing the parameter of interest. If equal to 
"total" then the overall F is used. The shortcut "int" works as well.}

\item{modObj}{Either an \code{lm} object, or a \code{summary.lm} object.}

\item{digits}{General number of digits to round to.}

\item{asEqn}{Should the statement be formatted as an equation?}

\item{dfDigits}{Digits to round degrees of freedom to.}

\item{pDigits}{Digits to round p-value to.}
}
\description{
Converts a model object to a summary object if necessary, validates
the parameter information, and grabs options if necessary.
}
