% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{rStrFromRcorr}
\alias{rStrFromRcorr}
\title{Constructs an APA formatted string for a correlation, but from lm().}
\usage{
rStrFromRcorr(rcorrObj, var1, var2, ...)
}
\arguments{
\item{...}{Options.}

\item{param}{The parameter of interest. Can be "int" as a shortcut 
for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}
}
\description{
Constructs an APA formatted string for a correlation, but from lm().
}
