% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResHelpUtility.R
\name{saveAllModelObjects}
\alias{saveAllModelObjects}
\title{Save All Model Objects to a File}
\usage{
saveAllModelObjects(file, envir = ".GlobalEnv", objClass = "lm", ...)
}
\arguments{
\item{file}{A string representing a filename to pass to \code{save()}.}

\item{envir}{Any way of specifying an environment in \code{ls()} and
\code{get()}.}

\item{objClass}{A character string of the class of object to be selected.}

\item{...}{Other options to pass to \code{save}}
}
\description{
This function is intended to save all objects of class "lm" to a specified
file. Note that by changing the \code{objClass} argument, it may be used to
save any kind of object to a file. You may also specify the environment in 
which to search for objects.
}
