% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{sdStr}
\alias{sdStr}
\title{Prints an APA formatted sd.}
\usage{
sdStr(v, na.rm = F, digits = NULL, asEqn = NULL)
}
\arguments{
\item{v}{Either a mean value or a vector. If a vector, then \code{sd()} 
is called on it.}

\item{na.rm}{If v is a vector, this is passed to \code{sd()}. 
Default: FALSE.}
}
\description{
Prints an APA formatted sd.
}
