% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{strNoLeadZero}
\alias{strNoLeadZero}
\title{Writes a decimal string with no leading zeros.}
\usage{
strNoLeadZero(val, digits = 3)
}
\arguments{
\item{val}{A numeric value.}
}
\value{
A character string.
}
\description{
Writes a decimal string with no leading zeros.
}
