% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{tStr}
\alias{tStr}
\title{Constructs an APA formatted string for a t-value.}
\usage{
tStr(param, modObj, digits = NULL, dfDigits = NULL)
}
\arguments{
\item{param}{The parameter of interest. Can be "int" as a shortcut 
for "(Intercept)".}

\item{modObj}{Either an \code{lm} or \code{summary.lm} (faster) object.}

\item{...}{Options.}
}
\description{
Constructs an APA formatted string for a t-value.
}
