% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{fpStr}
\alias{fpStr}
\title{Prints an APA formated string of the form:
\code{F(df1,df2),p[=,<]pval}.}
\usage{
fpStr(param, modObj, asEqn = NULL, digits = NULL, dfDigits = NULL)
}
\arguments{
\item{param}{The parameter of interest. The shortcut "int" is recognized
as well as the "total" term.}

\item{modObj}{The \code{lm} or \code{summary.lm} object.}

\item{asEqn}{Should output be a markdown equation?}

\item{...}{Other options.}
}
\description{
Prints an APA formated string of the form:
\code{F(df1,df2),p[=,<]pval}.
}
