% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{getOpts}
\alias{getOpts}
\title{A way to quickly get options for functions.}
\usage{
getOpts(arguments = NULL)
}
\arguments{
\item{arguments}{A character vector of arguments to access. Otherwise,
the function processes all available arguments.}
}
\description{
A way to quickly get options for functions.
}
\details{
If the
}
