% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resHelpStrFns.R
\name{tpStr}
\alias{tpStr}
\title{Prints an APA formated string of the form:
\code{t(df),p[=,<]pval}.}
\usage{
tpStr(param, modObj, ...)
}
\arguments{
\item{param}{The parameter of interest. The shortcut "int" is recognized.}

\item{modObj}{The \code{lm} or \code{summary.lm} object.}

\item{...}{Other options.}

\item{asEqn}{Should output be a markdown equation?}
}
\description{
Prints an APA formated string of the form:
\code{t(df),p[=,<]pval}.
}
